am62ax: mcasp: fix PRECISION.LOSS.CALL
The PRECISION.LOSS checker finds instances in which
an implicit cast to a smaller data type during a
function call can cause a loss of precision in data.

Changing the typedef from uint64_t to uint32_t will avoid the precision loss.

Fixes: SITSW-3457

+++ source/drivers/mcasp/v1/mcasp_dma.c
@@ -774,7 +774,7 @@ static void MCASP_udmaIsrTx(Udma_EventHandle eventHandle,
         MCASP_Transaction *txn = NULL;
         CSL_UdmapTR3 *pTr;
         uint32_t nextCandidate;
-        uint64_t txCnt = 0;
+        uint32_t txCnt = 0;
 
         nextCandidate = (object->lastFilled+1)%MCASP_TX_DMA_RING_ELEM_CNT;
 
@@ -787,7 +787,7 @@ static void MCASP_udmaIsrTx(Udma_EventHandle eventHandle,
 
             if(txn != object->reqQueueHandleTx)
             {
-                txCnt = (uint64_t)(txn->count*4);
+                txCnt = (txn->count*4);
 
                 if(txCnt < MCASP_DMA_L0_MAX_XFER_SIZE)
                 {
